requirements.txt
prompt_toolkit
rich
inquirer
python-dotenv
psycopg[binary]
networkx